+++ crates/ruma-client-api/src/r0/contact.rs
@@ -1,4 +1,0 @@
-//! Endpoints for account contact information.
-
-pub mod get_contacts;
-pub mod request_contact_verification_token;